requirements.txt
setuptools
pymysql
pyyaml
tarantool
